Tested the camera_ctrl service. Works fine!

+++ asctec_hl_comm/include/asctec_hl_comm/HL_interface.h
@@ -370,8 +370,8 @@ typedef struct
 __attribute__((packed))
 {
   /// desired roll and pitch angles for the camera
-  int16_t desired_cam_pitch;
-  int16_t desired_cam_roll;
+  int32_t desired_cam_pitch;
+  int32_t desired_cam_roll;
 }HLI_CAMERA;
 
 #endif
